Moved player registration card to Registration.svelte
